+++ test/chart/line_chart/line_chart_helper_test.dart
@@ -56,5 +56,31 @@ void main() {
       final result2 = LineChartHelper.calculateMaxAxisValues(lineBarsClone);
       expect(result1, result2);
     });
+
+    test('Test null spot 1', () {
+      final lineBars = [
+        LineChartBarData(spots: [
+          FlSpot.nullSpot,
+          FlSpot.nullSpot,
+          FlSpot.nullSpot,
+          FlSpot.nullSpot,
+        ])
+      ];
+      final result1 = LineChartHelper.calculateMaxAxisValues(lineBars);
+      expect(result1, LineChartMinMaxAxisValues(0, 0, 0, 0));
+    });
+
+    test('Test null spot 2', () {
+      final lineBars = [
+        LineChartBarData(spots: [
+          FlSpot.nullSpot,
+          const FlSpot(-1, 5),
+          FlSpot.nullSpot,
+          const FlSpot(4, -3),
+        ])
+      ];
+      final result1 = LineChartHelper.calculateMaxAxisValues(lineBars);
+      expect(result1, LineChartMinMaxAxisValues(-1, 4, -3, 5));
+    });
   });
 }